requirements.txt
flask==3.0.0
flask-login==0.6.3
flask-bcrypt==1.0.1
pymongo[srv]==4.6.1
python-dotenv==1.0.0
gunicorn==21.2.0